bug(api): PR changes
